Add RETURNING clause to UPDATE queries
This commit adds a RETURNING clause to UPDATE queries. Previously the
provided model was simply returned to the user, but this is wrong if
there are row-level triggers on the table. Adding RETURNING allows us to
return the correct model to the user.

Fix https://github.com/alevy/postgresql-orm/issues/13